Resolved a minor issue with WarningsDialogs
