+++ GameActivity.csproj
@@ -77,6 +77,7 @@
     <Compile Include="Common\HWiNFODumper.cs" />
     <Compile Include="GameActivity.cs" />
     <Compile Include="GameActivitySettings.cs" />
+    <Compile Include="Models\WarningData.cs" />
     <Compile Include="Services\PerfCounter.cs" />
     <Compile Include="Views\GameActivitySettingsView.xaml.cs">
       <DependentUpon>GameActivitySettingsView.xaml</DependentUpon>
